pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nonebot-plugin-tieba-monitor"
version = "0.1.3"
description = "NoneBot2 plugin for monitoring Tieba forums"
readme = "readme.md"
authors = [
    {name = "su-liu-guang", email = "wuhui0404@gmail.com"}
]
license = {text = "GNU AGPLv3"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Operating System :: OS Independent",
    "Framework :: Robot Framework :: Library",
]
dependencies = [
    "nonebot2>=2.3.0",
    "nonebot-adapter-onebot>=2.2.3",
    "nonebot-plugin-apscheduler>=0.3.0",
    "aiotieba>=1.4.0",
    "openai>=1.0.0",
    "nonebot_plugin_localstore>=0.72"

]
requires-python = ">=3.8"

[project.urls]
"Homepage" = "https://github.com/su-liu-guang/nonebot-plugin-tieba-monitor"
"Bug Tracker" = "https://github.com/su-liu-guang/nonebot-plugin-tieba-monitor/issues"

[tool.setuptools]
packages = ["nonebot_plugin_tieba_monitor"] 